Add support for optional lookAhead: parameter.
Allows masonry to "look ahead" x number of elements to find an element
that fits within the available columns better than the first one. This
will help fill in layouts where elements spanning several columns leave
large gaps in the layout if there are (widly) varying heights for the
elements.

This feature requires a version of outlayer.js that has the
_getItemIndexToLayOut function.