Update masonry.pkgd.js
Fix eventie helper used with requireJS :
To use Masonry with RequireJS, when we define a module, we can't define this module with an object because it's return null in other module definition like 'Outlayer'. To solve this error, we will pass a function to define() and return this object properly. This approach is claimed by requireJS documentation.